move periodic stuff Drivetrain -> PoseTracker
